+++ cmds/segment.go
@@ -1,4 +1,4 @@
-package main
+package cmds
 
 import (
 	lytics "github.com/lytics/go-lytics"
